requirements.txt
# Core
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.38.0
accelerate>=0.27.0
safetensors>=0.4.0

# Video I/O
decord>=0.6.0
av>=11.0.0

# Data / science
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
einops>=0.7.0

# Config / CLI
pyyaml>=6.0
tqdm>=4.66
rich>=13.7

# Logging / tracking
tensorboard>=2.15

# Visualization
matplotlib>=3.8
seaborn>=0.13
umap-learn>=0.5.5
opencv-python-headless>=4.9

# Dev
pytest>=8.0
